refactor the deflate, gz and zlib modules
